+++ crates/astria-bridge-withdrawer/src/bridge_withdrawer/submitter/mod.rs
@@ -15,6 +15,7 @@ use astria_core::{
         Action,
         TransactionBody,
     },
+    Protobuf as _,
 };
 use astria_eyre::eyre::{
     self,
@@ -142,6 +143,13 @@ impl Submitter {
             metrics,
             ..
         } = self;
+
+        if actions.is_empty() {
+            metrics.set_batch_total_settled_value(0);
+
+            return Ok(());
+        }
+
         // get nonce and make unsigned transaction
         let nonce = get_pending_nonce(
             sequencer_grpc_client.clone(),