+++ crates/astria-sequencer/src/action_handler/impls/bridge_unlock.rs
@@ -8,23 +8,24 @@ use astria_eyre::eyre::{
     Result,
     WrapErr as _,
 };
+use async_trait::async_trait;
 use cnidarium::StateWrite;
 
 use crate::{
-    accounts::action::{
+    action_handler::{
         check_transfer,
         execute_transfer,
+        ActionHandler,
     },
     address::StateReadExt as _,
-    app::ActionHandler,
     bridge::{
         StateReadExt as _,
-        StateWriteExt as _,
+        StateWriteExt,
     },
     transaction::StateReadExt as _,
 };
 
-#[async_trait::async_trait]
+#[async_trait]
 impl ActionHandler for BridgeUnlock {
     async fn check_stateless(&self) -> Result<()> {
         Ok(())
@@ -89,7 +90,6 @@ impl ActionHandler for BridgeUnlock {
 mod tests {
     use astria_core::{
         primitive::v1::{
-            asset,
             RollupId,
             TransactionId,
         },
@@ -102,8 +102,11 @@ mod tests {
 
     use crate::{
         accounts::StateWriteExt as _,
+        action_handler::{
+            impls::test_utils::test_asset,
+            ActionHandler as _,
+        },
         address::StateWriteExt as _,
-        app::ActionHandler as _,
         benchmark_and_test_utils::{
             assert_eyre_error,
             astria_address,
@@ -117,12 +120,8 @@ mod tests {
         },
     };
 
-    fn test_asset() -> asset::Denom {
-        "test".parse().unwrap()
-    }
-
     #[tokio::test]
-    async fn fails_if_bridge_account_has_no_withdrawer_address() {
+    async fn bridge_unlock_fails_if_bridge_account_has_no_withdrawer_address() {
         let storage = cnidarium::TempStorage::new().await.unwrap();
         let snapshot = storage.latest_snapshot();
         let mut state = StateDelta::new(snapshot);
@@ -130,7 +129,7 @@ mod tests {
         state.put_transaction_context(TransactionContext {
             address_bytes: [1; 20],
             transaction_id: TransactionId::new([0; 32]),
-            source_action_index: 0,
+            position_in_transaction: 0,
         });
         state.put_base_prefix(ASTRIA_PREFIX.to_string()).unwrap();
 
@@ -161,7 +160,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn fails_if_withdrawer_is_not_signer() {
+    async fn bridge_unlock_fails_if_withdrawer_is_not_signer() {
         let storage = cnidarium::TempStorage::new().await.unwrap();
         let snapshot = storage.latest_snapshot();
         let mut state = StateDelta::new(snapshot);
@@ -169,7 +168,7 @@ mod tests {
         state.put_transaction_context(TransactionContext {
             address_bytes: [1; 20],
             transaction_id: TransactionId::new([0; 32]),
-            source_action_index: 0,
+            position_in_transaction: 0,
         });
         state.put_base_prefix(ASTRIA_PREFIX.to_string()).unwrap();
 
@@ -204,7 +203,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn execute_with_duplicated_withdrawal_event_id() {
+    async fn bridge_unlock_executes_with_duplicated_withdrawal_event_id() {
         let storage = cnidarium::TempStorage::new().await.unwrap();
         let snapshot = storage.latest_snapshot();
         let mut state = StateDelta::new(snapshot);
@@ -213,7 +212,7 @@ mod tests {
         state.put_transaction_context(TransactionContext {
             address_bytes: bridge_address.bytes(),
             transaction_id: TransactionId::new([0; 32]),
-            source_action_index: 0,
+            position_in_transaction: 0,
         });
         state.put_base_prefix(ASTRIA_PREFIX.to_string()).unwrap();